Rename tasklist.component.ts to task-list.component.ts Rename tasklist.component.md to task-list.component.md
