+++ ng2-components/ng2-activiti-tasklist/src/components/task-list.component.css
